Inherit from ActionController::Base
To avoid collision of the Authentication modules of the
different authentication plugins, we isolate completely
our application. Inheriting from AC::Base instead of
ApplicationController.

As a bonus we don't have to skip application filters.

A small refactor will be made to remove current code
duplication in Admin::MasterController and TypusController.

Thanks to Andrey Deryabin (http://github.com/aderyabin) for
the suggestion.