Removed everything related to RichText editors.
This adds some agnosticism to the application. I will
document how to add a RichText editor without many effort.

Until today Typus was using rails-ckeditor plugin which
is nice but has a bug that leaves the application open to
the world so anyone would be able to send data through your
form if not action is taken.