Detaching a file (paperclip/dragonfly) refactored.
Let's say for example you have a model Asset with `caption` and `file`
as attributes.

Lets imagine you have uploaded a file with no `caption`. And in the future
you want validate the presence of `caption`.

With the previous implementation we were only getting a "File can't be
removed." error message, but we didn't know which was the exact reason
why the file couldn't be removed. This refactor fixes this behavior.