Removed :read_only option.
Read only option was one of those things I added for a project an never
have used again. It was interesting to have but useless because there
was no mass-assignment protection. With Rails 3.1 we can use mass-assignment
protection based on Roles which is something really usefull.

If you want to protect an attribute to be edited, so it's "read only" you
need to use `mattr_protected`. Eg.

    class Post < ActiveRecord::Base
      attr_protected :permalink
    end

This will protect editing the permalink for the posts. Interesting right?
If you are logged as `admin` you can still edit it.