Updated how tabs work. Filters by app instead of going to model.
Previously when clicking on an application tabs users were redirected
to the first model of that application.

Now clicking on an application tab redirects the user to a list of
models inside that application, consider it a filter of the applications
listed on the dashboard.

This is how Django Works and this is how Typus should work too.

Interesting points:

- In the future we will be able to define views for each application.
  Imagine a "Dashboard" for every application.

- Fixes problem with nested models when using more than 2 levels.