Typus v3.1.0
Finally after a few months of hard work and a big refactor of Typus internals
to make it cleaner, nicer and hopefully faster I'm happy to release it as
always under the MIT License.

There are still some bugs and things to polish, but I need to release it
because people, specially clients are waiting to start using it.