Hack to be able to disable "Chosen".
What happens with "Chosen"?

"Chosen" is a really good plugin but at this moment it doesn't handle
remote data loading, we can work without this feature but if there are
more than 250 records the search feature is painfully slow, so now we
have an option which will allow developers to disable "Chosen" when
needed.

I know, this is a super hack, but I'm don't have time to work on this
and prefer to wait for "Chosen" to support this.