Remove #available_locales, the last method of Typus::I18n.
The information is now accessible in Typus.available_locales. Right
now it is hard-coded to English only, but this will have to be made smarter
when the new gem typus-i18n is also loaded. (it would be best to figure out 
dynamically which translations are available for the keyspace xx.typus.*“).